- ManagerDaoImpl#delete(Gebruiker) correctly handles the case where a user has no database backing. - Created an abstract LDAP base test class. - Added a test for deleting users with database backing.
