+++ managerDB/src/test/java/nl/ipo/cds/dao/ManagerDaoAuthenticationProviderTest.java
@@ -1,9 +1,12 @@
 package nl.ipo.cds.dao;
 
 import static org.junit.Assert.*;
+import nl.ipo.cds.categories.IntegrationTests;
 
 import org.junit.Test;
+import org.junit.experimental.categories.Category;
 
+@Category(IntegrationTests.class)
 public class ManagerDaoAuthenticationProviderTest extends BaseManagerDaoTest {
 
 	@Test
